+++ src/MicroElements.Metadata/Metadata/Reflection.cs
@@ -8,6 +8,9 @@ using System.Reflection;
 
 namespace MicroElements.Metadata
 {
+    /// <summary>
+    /// Metadata reflection helpers.
+    /// </summary>
     public static class Reflection
     {
         /// <summary>
@@ -15,7 +18,7 @@ namespace MicroElements.Metadata
         /// </summary>
         /// <param name="type">Type that holds static fields and properties.</param>
         /// <param name="prefix">Prefix to add to begin of the property name.</param>
-        public static void AutoCreateStaticProperties(this Type type, string prefix = null)
+        public static void AutoCreateStaticProperties(this Type type, string? prefix = null)
         {
             type
                 .GetFields(BindingFlags.Static | BindingFlags.Public)
@@ -41,6 +44,30 @@ namespace MicroElements.Metadata
                     IProperty property = Property.Create(propertyType, propertyName);
                     propertyInfo.SetValue(null, property);
                 });
+        }
+
+        public static IReadOnlyCollection<IProperty> GetStaticProperties(this Type type, bool autoCreateStaticProperties = true, string? prefix = null)
+        {
+            var fromFields = type
+                .GetFields(BindingFlags.Static | BindingFlags.Public)
+                .Where(fieldInfo => typeof(IProperty).IsAssignableFrom(fieldInfo.FieldType))
+                .Select(fieldInfo => fieldInfo.GetValue(null))
+                .Cast<IProperty>();
+
+            var fromProperties = type
+                .GetProperties(BindingFlags.Static | BindingFlags.Public)
+                .Where(propertyInfo => typeof(IProperty).IsAssignableFrom(propertyInfo.PropertyType))
+                .Select(propertyInfo => propertyInfo.GetValue(null))
+                .Cast<IProperty>();
+
+            IReadOnlyCollection<IProperty> properties = fromFields.Concat(fromProperties).ToArray();
+            if (autoCreateStaticProperties && properties.Any(property => property == null))
+            {
+                AutoCreateStaticProperties(type, prefix);
+                properties = GetStaticProperties(type, autoCreateStaticProperties: false);
+            }
+
+            return properties;
         }
     }