+++ src/MicroElements.Metadata/Metadata/Schema/IMetadata.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace MicroElements.Metadata.Schema
+{
+    /// <summary>
+    /// Marker interface for metadata types.
+    /// </summary>
+    public interface IMetadata
+    {
+    }
+}